Web Scrapping Code, PPT, Text File, Read Me word document
Web scraping is the process of collecting and parsing raw data from the web page. Web Scraping involves fetching it and extracting from it. Fetching is the downloading of a page (which a browser does when a user views a page). Therefore, web crawling is a main component of web scraping, to fetch pages for later processing. Once fetched, extraction can take place. The content of a page may be parsed, searched and reformatted, and its data copied into a spreadsheet or loaded into a database.